Move challenge42 to a new package for consistency
