+++ src/main/resources/wrong-secrets-configuration.yaml
@@ -688,7 +688,7 @@ configurations:
     - name: Challenge 42
       short-name: "challenge-42"
       sources:
-        - class-name: "org.owasp.wrongsecrets.challenges.docker.Challenge42"
+        - class-name: "org.owasp.wrongsecrets.challenges.docker.challenge42.Challenge42"
           explanation: "explanations/challenge42.adoc"
           hint: "explanations/challenge42_hint.adoc"
           reason: "explanations/challenge42_reason.adoc"
